chore: extract sessionPublicContext
